feat: add shell hook
This adds the necessary shell hook that allows you to set breakpoints in
the activation script. The way it works is that you sprinkle calls to
`fdb_tracepoint <name>` in the script at locations that you think you
might want to stop and inspect the program state. When you want to set a
breakpoint at this location, you run the program with the
`FLOX_DBG_TRACEPOINT=<value>` variable set, where `<value>` can be the
name of a specific tracepoint, `next`, or `all`. When the program hits
this tracepoint, the function will check whether the variable is set,
and if it's set whether it matches one of these predefined values or the
name of the tracepoint itself. In this scenario we launch the debugger
and source its output once the debugger exits. This allows the debugger
to influence control flow after it has exited e.g. by unsetting the
variable, or leaving it set to `next` or `all`.